refactor(floci): regroup UI and DuckDB under services.floci
Move the UI service classes (FlociUiManager, UiPages) into services.floci.ui and the DuckDB helper (FlociDuckClient, FlociDuckManager) into services.floci.duck, with tests mirrored. Pure package move: importers updated, no behavior change.
