+++ src/test/java/io/github/hectorvent/floci/lifecycle/EmulatorLifecycleTest.java
@@ -65,7 +65,7 @@ class EmulatorLifecycleTest {
     @Mock private PipesService pipesService;
     @Mock private Ec2MetadataServer ec2MetadataServer;
     @Mock private EcrRegistryManager ecrRegistryManager;
-    @Mock private io.github.hectorvent.floci.services.ui.FlociUiManager flociUiManager;
+    @Mock private io.github.hectorvent.floci.services.floci.ui.FlociUiManager flociUiManager;
     @Mock private InitLifecycleState initLifecycleState;
     @Mock private EmulatorConfig.TlsConfig tlsConfig;
