+++ src/test/java/io/github/hectorvent/floci/services/cur/CurEmissionIntegrationTest.java
@@ -1,6 +1,6 @@
 package io.github.hectorvent.floci.services.cur;
 
-import io.github.hectorvent.floci.services.floci.FlociDuckClient;
+import io.github.hectorvent.floci.services.floci.duck.FlociDuckClient;
 import io.github.hectorvent.floci.services.s3.S3Service;
 import io.github.hectorvent.floci.testing.RestAssuredJsonUtils;
 import io.quarkus.test.junit.QuarkusTest;
